% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demultiplexer.R
\name{combinatorial_demultiplex}
\alias{combinatorial_demultiplex}
\title{Combinatorial demultiplexer}
\usage{
combinatorial_demultiplex(
  sequences,
  barcodes,
  sequence_annotation,
  segment_lengths
)
}
\arguments{
\item{sequences}{A \code{\link{XStringSet}} object, the sequences to be demultiplexed
All the sequences must be named}

\item{barcodes}{A named list of \code{\link{XStringSet}} objects, the barcodes
to be used for demultiplexing. All of the barcodes in each \code{\link{XStringSet}} must
have the same length as specified by the \code{segment_lengths} argument}

\item{segment_lengths}{Integer vector with the same length
as \code{segments}, lengths of the segments.
Up to one of the non-barcode segments can have its length
set to \code{NA} which means
it is considered a variadic length segment}

\item{segments}{Character vector showing the segments of the
sequences from 5' end to 3' end. The code applied is as follows:
\itemize{
\item \code{'A'}: Adapter, it is trimmed and ignored
\item \code{'B'}: Barcode, used for demultiplexing
\item \code{'P'}: Payload, sequence to be kept after trimming
and demultiplexering (e.g. cDNA or UMI).
}}
}
\value{
A list with the following elements:
\itemize{
\item \code{assigned_barcodes}: A \code{character} matrix with
the names of the assigned barcodes as elements. The rows correspond to the
sequences and the columns to the barcode segments
\item \code{mismatches}: An \code{integer} matrix with the number of mismatches
between the assigned barcodes and the sequences. The rows correspond to the
sequences and the columns to the barcode segments.
\item \code{payload}: A \code{\link{XStringSet}} object with the payload sequences
}
}
\description{
This function performs combinatorial demultiplexing and trimming
of sequences. The sequences have a structure as defined by the argument
\code{segments} with corresponding lengths in \code{segment_lengths}.
As trimming and extraction can take place from either end, a single
middle segment can be variadic in length. There are three types of segments:
Adapter, Barcode and Payload. The adapter is trimmed and ignored, the barcode
is used for demultiplexing, and the payload is kept after trimming and demultiplexing
and returned from the function. If there are multiple payload segments, they
concatented in the output sequences. The barcodes can be positioned at
either end of the sequences,
but no barcode can (for obvious reasons) be variadic in length.
}
