% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_operations.R
\name{demultiplex_pipeline}
\alias{demultiplex_pipeline}
\title{Run pipeline for demultiplexing sequences}
\usage{
demultiplex_pipeline(
  input_file,
  reverse_file,
  output_table_file,
  barcode_files,
  sequence_annotation,
  segment_lengths,
  allowed_mismatches,
  trimmed_output_file,
  reverse_output_file
)
}
\arguments{
\item{reverse_file}{The file path to the reverse reads
corresponding to \code{input_file}. The contents of the reverse reads
will be filtered according to the same barcode assignments as the forward reads.}

\item{reverse_output_table_file}{If \code{reverse_read} the file path to the table}
}
\description{
Run pipeline for demultiplexing sequences
}
