% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_operations.R
\name{file_combinatorial_demultiplex}
\alias{file_combinatorial_demultiplex}
\title{Demultiplex combinatorial barcodes}
\usage{
file_combinatorial_demultiplex(
  input_file,
  output_table_file,
  barcode_files,
  sequence_annotation,
  segment_lengths,
  allowed_mismatches = 0L,
  trimmed_output_file = NULL
)
}
\arguments{
\item{input_file}{The file path to the fastq sequences to be demultiplexed,}

\item{output_table_file}{The file path to the table
showing the barcode assignments of each sequences}

\item{barcode_files}{Named character vector,
the file paths to the barcodes to be demultiplexed, arranged in the same
order as in the sequence annotation}

\item{segment_lengths}{Integer vector with the same length
as \code{segments}, lengths of the segments.
Up to one of the non-barcode segments can have its length
set to \code{NA} which means
it is considered a variadic length segment}

\item{allowed_mismatches}{Maximum Hamming distance from sequences
to barcodes, either a single integer or an integer vector with one number
per barcode file}

\item{trimmed_output_file}{The file path to the fastq files
containing the  sequences of the payload after adapters
and barcodes have been trimmed. If \code{NULL}, no such file will be created}

\item{segments}{Character vector showing the segments of the
sequences from 5' end to 3' end. The code applied is as follows:
\itemize{
\item \code{'A'}: Adapter, it is trimmed and ignored
\item \code{'B'}: Barcode, used for demultiplexing
\item \code{'P'}: Payload, sequence to be kept after trimming
and demultiplexering (e.g. cDNA).
}}
}
\value{
The output from \code{\link{filter_demultiplex_res}}. In addition,
the function will write to the file paths specified in \code{output_table_file}
and (optionally) \code{trimmed_output_file}.
}
\description{
Demultiplex combinatorial barcodes
}
\details{
The barcode table will be tab separated with the first row showing the
barcode name.
names.
}
