% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_filter.R
\name{filter_demultiplex_res}
\alias{filter_demultiplex_res}
\title{Filter demultiplexed reads based on the number of mismatches}
\usage{
filter_demultiplex_res(demultiplex_res, allowed_mismatches)
}
\arguments{
\item{demultiplex_res}{Unprocessed output from
\code{\link{combinatorial_demultiplex}}}

\item{allowed_mismatches}{Integer vector of length one or the same length
as the number of barcode segments. The maximum Hamming distance from including
a read in the output}
}
\value{
A list
}
\description{
Filter demultiplexed reads based on the number of mismatches
}
