% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode_frequency.R
\name{frequency_plot}
\alias{frequency_plot}
\alias{knee_plot}
\title{Diagnostic plots from demultiplexing}
\usage{
frequency_plot(
  frequency_table,
  cutoff = NULL,
  type = "histogram",
  log_scale_x = TRUE,
  log_scale_y = FALSE
)

knee_plot(frequency_table, cutoff = NULL)
}
\arguments{
\item{frequency_table}{The frequency table
from \code{\link{create_frequency_table}}}

\item{cutoff}{Optional scalar numeric, the
x-coordinate for drawing a vertical dashed line
in the plots in order to indicate the cutoff. Please note that this argument
is interpreted literally, meaning the in order to correctly display
the same cutoff on both type of plots,
the cutoff values has to be transformed. In order to safely convert between
the two types of cutoffs, use the functions
\code{\link{bc_to_frequency_cutoff}}
and \code{\link{frequency_to_bc_cutoff}}.}

\item{type}{The type of frequency plot to make, either \code{"histogram"}
or \code{"density"}}

\item{log_scale_x}{Logical: Should a log scale be applied to the x-axis of the
frequency plot?}

\item{log_scale_y}{Logical: Should a log scale be applied to the y-axis of the
frequency plot?}
}
\value{
A \code{\link{ggplot}} which can be displayed immediately or
further modified
}
\description{
Diagnostic plots for determining the effect of the barcode cutoff.
\code{frequency_plot()} shows a histogram or distribution plot of the
number of reads for each barcode
combination, whereas \code{knee_plot()} shows the cumulative fraction of reads
ranked by the frequency of the barcode combinations in descending order.
}
\seealso{
\code{\link[=bc_to_frequency_cutoff]{bc_to_frequency_cutoff()}} \code{\link[=frequency_to_bc_cutoff]{frequency_to_bc_cutoff()}}
}
