% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff_select.R
\name{interactive_bc_cutoff}
\alias{interactive_bc_cutoff}
\title{Interactive Shiny application for determining cutoff from knee plot}
\usage{
interactive_bc_cutoff(
  frequency_table,
  viewer = "dialog",
  viewer_height = 800,
  viewer_width = 1200
)
}
\arguments{
\item{frequency_table}{}
}
\description{
Interactive Shiny application for determining cutoff from knee plot
}
