% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posDemux-package.R
\docType{package}
\name{posDemux}
\alias{posDemux-package}
\alias{posDemux}
\title{posDemux: Positional combinatorial sequence demultiplexer}
\description{
More about what it does (maybe more than one line).
}
\author{
Jakob Peder Pettersen \href{mailto:jakobpeder.pettersen@gmail.com}{jakobpeder.pettersen@gmail.com}
}
