% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_filter.R
\name{print.demultiplex_filter_summary}
\alias{print.demultiplex_filter_summary}
\title{Diagnostic demultiplexing results}
\usage{
\method{print}{demultiplex_filter_summary}(x, ...)
}
\arguments{
\item{x}{The field \code{filter_summary}  from \code{\link{filter_demultiplex_res}}}

\item{...}{Ignored}
}
\value{
Its input, invisibly.
}
\description{
Prints diagnostic information about the results of demultiplexing
and subsequent filtering, including the results per barcode set.
}
