% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_filter.R
\name{print.filter_summary}
\alias{print.filter_summary}
\title{Diagnostic demultiplexing results}
\usage{
\method{print}{filter_summary}(x)
}
\arguments{
\item{x}{The field \code{filter_summary}  from \code{\link{filter_demultiplex_res}}}
}
\value{
Its input, invisibly
}
\description{
Diagnostic demultiplexing results
}
