% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/demultiplexer.R
\name{combinatorial_demultiplex}
\alias{combinatorial_demultiplex}
\title{Combinatorial demultiplexer}
\usage{
combinatorial_demultiplex(sequences, barcodes, segments, segment_lengths)
}
\arguments{
\item{sequences}{A \code{\link{XStringSet}} object, the sequences to be demultiplexed}

\item{barcodes}{A list of \code{\link{XStringSet}} objects, the barcodes
to be used for demultiplexing. All of the barcodes in each \code{\link{XStringSet}} must
have the same length as specified by the \code{segment_lengths} argument and be named.
For computational reasons, the maximum possible length of an individual barcode
is 127.}

\item{segments}{Character vector showing the segments of the
sequences from 5' end to 3' end. The code applied is as follows:
\itemize{
\item \code{'A'}: Adapter, it is trimmed and ignored
\item \code{'B'}: Barcode, used for demultiplexing
\item \code{'P'}: Payload, sequence to be kept after trimming
and demultiplexering (e.g. cDNA or UMI).
}}

\item{segment_lengths}{Integer vector with the same length
as \code{segments}, lengths of the segments.
Up to one of the non-barcode segments can have its length
set to \code{NA} which means
it is considered a variadic length segment}
}
\value{
A list with the following elements:
\itemize{
\item \code{assigned_barcodes}: A \code{character} matrix with
the names of the assigned barcodes as elements. The rows correspond to the
sequences and the columns to the barcode segments
\item \code{mismatches}: An \code{integer} matrix with the number of mismatches
between the assigned barcodes and the sequences. The rows correspond to the
sequences and the columns to the barcode segments.
\item \code{payload}: A \code{\link{XStringSet}} object with the payload sequences
\item \code{barcodes}: The \code{barcodes} argument passed into the function.
It is included in order to ease downstream processing.
}
}
\description{
This function performs combinatorial demultiplexing and trimming
of sequences. The sequences have a structure as defined by the argument
\code{segments} with corresponding lengths in \code{segment_lengths}.
As trimming and extraction can take place from either end, a single
middle segment can be variadic in length. There are three types of segments:
Adapter, Barcode and Payload. The adapter is trimmed and ignored, the barcode
is used for demultiplexing, and the payload is kept after trimming and demultiplexing
and returned from the function. If there are multiple payload segments, they
concatented in the output sequences. The barcodes can be positioned at
either end of the sequences,
but no barcode can (for obvious reasons) be variadic in length.
}
\details{
If there are two barcodes both having the minimum number of mismatches
the first one will be selected. If is therefore important to choose the
error tolerance to be equal or less than the redunancy of the barcodes.
}
