% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{create_frequency_table}
\alias{create_frequency_table}
\title{Create a frequency table of the observed barcode combinations}
\usage{
create_frequency_table(filter_results)
}
\arguments{
\item{filter_results}{The output from \code{\link{filter_demultiplex_res}}}
}
\value{
A data frame where each row correspond to a unique observed
barcode combination. The rows are sorted in descending order of frequency.
The first columns are specify the barcode assignment and the last columns were the
following:
\itemize{
\item \code{frequency}: The number of reads with the barcode combination
\item \code{cumulative_frequency}: The cumulative frequency of the barcode combination
\item \code{fractiion}: The fraction of reads with the barcode combination
\item \code{cumulative_fraction}: The cumulative fraction of the barcode combination
}
\description{
Create a frequency table of the observed barcode combinations
}
