% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/barcode_frequency.R
\name{create_frequency_table}
\alias{create_frequency_table}
\title{Frequency table}
\usage{
create_frequency_table(assigned_barcode)
}
\arguments{
\item{assigned_barcode}{A character or integer matrix, corresponding to the field
\code{assigned_barcode} from \code{\link{combinatorial_demultiplex}}
or the field \code{demultiplex_res$assigned_barcode}
from \code{\link{combinatorial_demultiplex}}}
}
\value{
A data frame where each row correspond to a unique observed
barcode combination. The rows are sorted in descending order of frequency.
The first columns are specify the barcode assignment and the last columns were the
following:
\itemize{
\item \code{frequency}: The number of reads with the barcode combination
\item \code{cumulative_frequency}: The cumulative frequency of the barcode combination
\item \code{fractiion}: The fraction of reads with the barcode combination
\item \code{cumulative_fraction}: The cumulative fraction of the barcode combination
}
}
\description{
Creates a sorted frequency table of each of the observed
barcode combinations. This function is indended to be used after running
\code{\link{filter_demultiplex_res}} and before creating frequency plots,
knee plots or selecting the number of barcodes to include.
}
