% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_filter.R
\name{filter_demultiplex_res}
\alias{filter_demultiplex_res}
\title{Filter demultiplexed reads}
\usage{
filter_demultiplex_res(demultiplex_res, allowed_mismatches)
}
\arguments{
\item{demultiplex_res}{Unprocessed output from
\code{\link{combinatorial_demultiplex}}}

\item{allowed_mismatches}{Integer vector of length one or the same length
as the number of barcode segments, the maximum Hamming distance from including
a read in the output.}
}
\value{
A list with the following elements:
\itemize{
\item \code{demultiplex_res}: The contents of \code{demultiplex_res} with
the sequences filtered.
\item \code{retained}: Logical vector with the same length as
the number of reads in the input. \code{TRUE} if the corresponding read
is retained.
Useful for future filtering of paired-end reads.
\item \code{summary_res}: Result of \code{\link{create_summary_res}} called
on the results of filtering.
}
}
\description{
Filters the demultiplexed reads from \code{\link{combinatorial_demultiplex}}
such that any read exceeding the number of allowed mismatches for any
of the barcodes is removed. The function gives diagnostic information
on the number of reads removed per barcode and the total number
of reads removed.
}
\details{
The value of \code{n_removed} does not in general equal the sum of
\code{n_removed_per_barcode} since a read can have too many mismatches
with multiple barcodes.
}
