% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{filter_summary}
\alias{filter_summary}
\title{Diagnostic demultiplexing results}
\usage{
filter_summary(filter_results)
}
\arguments{
\item{filter_results}{The output from \code{\link{filter_demultiplex_res}}}
}
\description{
Diagnostic demultiplexing results
}
