% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cutoff_select.R
\name{interactive_bc_cutoff}
\alias{interactive_bc_cutoff}
\title{Interactive cutoff selection}
\usage{
interactive_bc_cutoff(frequency_table)
}
\arguments{
\item{frequency_table}{The frequency table
from \code{\link{create_frequency_table}}}
}
\value{
Integer scalar with the selected cutoff returned
invisibly
}
\description{
Launches an interactive Shiny application for
determining cutoff from knee plot and barcode frequency plot. The user
will select the appropriate number of barcode combinations
(i.e. distinct cells) to keep for further analysis. Usually, this is done
by aiming for the "knee" of the knee plot in order to keep most reads,
while at the same time removing barcode combinations which are either
artifacts or broken cells.
}
