% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/posDemux-package.R
\name{posDemux}
\alias{posDemux}
\title{Utilities for segmented sequences with combinatorial barcoding}
\description{
Provides tools for handling reads with combinatorial barcodes, and
multiple adapter regions. This includes utilities for demultiplexing,
sequence segmenting and filtering. The package is intended to work with
single-cell RNA-seq data with multiple barcoding
(e.g. SPLiT-seq and PETRI-seq).
}
\author{
Jakob Peder Pettersen \href{mailto:jakobpeder.pettersen@gmail.com}{jakobpeder.pettersen@gmail.com}
}
