% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_summary.R
\name{print.demultiplex_filter_summary}
\alias{print.demultiplex_filter_summary}
\title{Prints diagnostic demultiplexing results}
\usage{
\method{print}{demultiplex_filter_summary}(x, ...)
}
\arguments{
\item{x}{An object of class \code{demultiplex_filter_summary} from
\code{\link{create_summary_res}}}

\item{...}{Ignored}
}
\value{
Its input, invisibly.
}
\description{
Prints diagnostic information about the results of demultiplexing
and subsequent filtering, including the results per barcode set.
}
